impr: move force-unlock modal to its own file and add test
