+++ libs/domains/services/feature/src/index.ts
@@ -55,3 +55,4 @@ export * from './lib/service-icon/service-icon'
 export * from './lib/service-access-modal/service-access-modal'
 export * from './lib/service-deployment-list/service-deployment-list'
 export * from './lib/pod-details/pod-details'
+export * from './lib/force-unlock-modal/force-unlock-modal'
